python/django-runscript/requirements.txt
Django>=5.1.2
django-extensions
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-distro
opentelemetry-instrumentation-botocore
circus
requests
boto3
certifi